backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests
python-dotenv
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.22.0
h11
httpcore==1.0.8
httpx==0.28.1
idna==3.10
joblib==1.4.2
lxml==5.3.2
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
phi==0.6.7
phidata==2.7.10
pillow==11.2.1
primp==0.14.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
tldextract==5.2.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
